Questions page

+++ ComedyHackDay/app/src/main/java/waggoner/com/comedyhackday/Answer.java
@@ -4,11 +4,11 @@ package waggoner.com.comedyhackday;
  * Created by nathanielwaggoner on 10/24/15.
  */
 public class Answer {
-    public Answer(String text, int fucks_given) {
+    public Answer(String text, float fucks_given) {
         this.text = text;
         this.fucks_given = fucks_given;
     }
 
     public String text;
-    public int fucks_given;
+    public float fucks_given;
 }
